Feature/make the service for create a category front

+++ client/app/services/adminService.ts
@@ -1,5 +1,5 @@
 import type { User } from "../types/userlist";
-
+import { CookieHelper } from "../utils/cookieHelper";
 export const adminService = {
   async authenticated(token: string) {
     try {
@@ -49,10 +49,18 @@ export const adminService = {
   },
 
   deleteCategory: async (idCategory: number) => {
+    const token = CookieHelper.getToken("AccesToken");
+    if (!token) {
+      return { success: false, error: "Token manquant" };
+    }
+
     try {
       const res = await fetch("http://localhost:3000/api/category/", {
         method: "DELETE",
-        headers: { "Content-Type": "application/json" },
+        headers: {
+          Authorization: `Bearer ${token}`,
+          "Content-Type": "application/json",
+        },
         body: JSON.stringify({ id: idCategory }),
       });
       if (!res.ok) {
@@ -91,11 +99,47 @@ export const adminService = {
     }
   },
   modifyCategory: async (name: string, description: string, id: number) => {
+    const token = CookieHelper.getToken("AccesToken");
+    if (!token) {
+      return { success: false, error: "Token manquant" };
+    }
     try {
       const res = await fetch("http://localhost:3000/api/category/", {
         method: "PUT",
-        headers: { "Content-Type": "application/json" },
+        headers: {
+          Authorization: `Bearer ${token}`,
+          "Content-Type": "application/json",
+        },
         body: JSON.stringify({ name, description, id }),
+      });
+      if (!res.ok) {
+        if (res.status === 401) {
+          throw new Error("Token invalide ou expiré");
+        } else if (res.status === 403) {
+          throw new Error("Droits administrateur requis");
+        }
+      }
+      const data = await res.json();
+      return { success: true, data };
+    } catch (err) {
+      console.error(`erreur: ${err}`);
+      return { success: false, error: "Erreur réseau ou serveur" };
+    }
+  },
+
+  createCategory: async (name: string, description: string) => {
+    const token = CookieHelper.getToken("AccesToken");
+    if (!token) {
+      return { success: false, error: "Token manquant" };
+    }
+    try {
+      const res = await fetch("http://localhost:3000/api/category/", {
+        method: "POST",
+        headers: {
+          Authorization: `Bearer ${token}`,
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify({ name, description }),
       });
       if (!res.ok) {
         if (res.status === 401) {
